refactor: restructure project with MVC architecture
